requirements.txt
tensorflow>=2.13.0
Pillow>=10.0.0
flask==2.3.3
flask-cors==4.0.0
scipy>=1.10.1
numpy>=1.23.5
pandas>=1.5.3
scikit-learn>=1.2.0
albumentations>=1.3.0